requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.12.0=pyhd8ed1ab_0
aiohttp=3.7.4=py37h27cfd23_1
argon2-cffi=20.1.0=py37h7b6447c_1
async-timeout=3.0.1=py_1000
async_generator=1.10=py37h28b3542_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.2.1=py_0
blinker=1.4=py_1
brotlipy=0.7.0=py37hb5d75c8_1001
bzip2=1.0.8=h516909a_3
c-ares=1.17.1=h36c2ea0_0
ca-certificates=2020.12.5=ha878542_0
cachetools=4.2.1=pyhd8ed1ab_0
cairo=1.14.12=h8948797_3
certifi=2020.12.5=py37h89c1867_1
cffi=1.14.4=py37h261ae71_0
chainer=7.7.0=pypi_0
chainercv=0.13.1=pypi_0
chardet=3.0.4=py37he5f6b98_1008
click=7.1.2=pyh9f0ad1d_0
cloudpickle=1.6.0=py_0
cryptography=3.4.6=py37h5d9358c_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.10.0=py_2
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2.30.0=py_0
dataclasses=0.6=pypi_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py37_0
expat=2.2.10=he6710b0_2
ffmpeg=4.0=hcdf2ecd_0
filelock=3.0.12=pypi_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.2.1=h58526e2_0
freetype=2.10.4=h5ab3b9f_0
future=0.18.2=pypi_0
glib=2.66.1=h92f7085_0
google-auth=1.24.0=pyhd3deb0d_0
google-auth-oauthlib=0.4.1=py_2
graphite2=1.3.13=h58526e2_1001
grpcio=1.33.2=py37haffed2e_2
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
h5py=2.8.0=py37h989c5e5_3
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hc401514_3
icu=58.2=he6710b0_3
idna=2.10=pyh9f0ad1d_0
imageio=2.9.0=py_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.19.0=py37hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.5.1=py_1
jasper=2.0.14=h07fcdf6_1
jedi=0.17.2=py37h06a4308_1
jinja2=2.11.2=py_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py_2
jupyter=1.0.0=py37_7
jupyter_client=6.1.7=py_0
jupyter_console=6.2.0=py_0
jupyter_core=4.7.0=py37h06a4308_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.3.1=py37hc928c03_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran=3.0.0=1
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=he1b5a44_1001
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.14.0=h8c45485_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.2=he1b5a44_3
markdown=3.3.4=pyhd8ed1ab_0
markupsafe=1.1.1=py37h14c3975_1
matplotlib=3.3.3=py37h89c1867_0
matplotlib-base=3.3.3=py37h4f6019d_0
mistune=0.8.4=py37h14c3975_1001
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
multidict=5.1.0=py37h27cfd23_2
nbclient=0.5.1=py_0
nbconvert=6.0.7=py37_0
nbformat=5.0.8=py_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.3=pyhd3eb1b0_0
networkx=2.5=py_0
ninja=1.10.2=py37hff7bd54_0
notebook=6.1.4=py37_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
oauthlib=3.0.1=py_0
olefile=0.46=pyh9f0ad1d_1
opencv=3.4.2=py37h6fd60c2_1
openssl=1.1.1j=h27cfd23_0
packaging=20.7=pyhd3eb1b0_0
pandoc=2.11=hb0f4dca_0
pandocfilters=1.4.3=py37h06a4308_1
parso=0.7.0=py_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.0.1=py37he98fc37_0
pip=20.3=py37h06a4308_0
pixman=0.40.0=h36c2ea0_0
prometheus_client=0.9.0=pyhd3eb1b0_0
prompt-toolkit=3.0.8=py_0
prompt_toolkit=3.0.8=0
protobuf=3.14.0=pypi_0
ptflops=0.6.4=pypi_0
ptyprocess=0.6.0=pyhd3eb1b0_2
py-opencv=3.4.2=py37hb342d67_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=py_2
pygments=2.7.2=pyhd3eb1b0_0
pyjwt=2.0.1=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.17.3=py37h7b6447c_0
pysocks=1.7.1=py37h89c1867_3
python=3.7.9=h7579374_0
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.7.0=py3.7_cuda11.0.221_cudnn8.0.3_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.3.1=py37h7b6447c_1
pyzed=3.4=pypi_0
pyzmq=20.0.0=py37h2531618_1
qt=5.9.7=h5867ecd_1
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
readline=8.0=h7b6447c_0
requests=2.25.1=pyhd3deb0d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.7.2=pyh44b312d_0
scikit-image=0.17.2=py37hdf5156a_0
scipy=1.5.2=py37h0b6359f_0
send2trash=1.5.0=py37_0
setuptools=50.3.2=py37h06a4308_2
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py37h06a4308_0
sqlite=3.33.0=h62c20be_0
tensorboard=2.4.1=pyhd8ed1ab_0
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
tensorboardx=2.1=py_0
terminado=0.9.1=py37_0
testpath=0.4.4=py_0
tifffile=2020.10.1=py37hdd07704_2
tk=8.6.10=hbc83047_0
toolz=0.11.1=py_0
torchvision=0.8.1=py37_cu110
tornado=6.1=py37h27cfd23_0
tqdm=4.54.0=pyhd8ed1ab_0
traitlets=5.0.5=py_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.4=pyhd8ed1ab_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.36.0=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py37_0
xorg-fixesproto=5.0=h14c3975_1002
xorg-inputproto=2.3.2=h14c3975_1002
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libx11=1.6.12=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxfixes=5.0.3=h516909a_1004
xorg-libxi=1.7.10=h516909a_0
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=py37h4abf009_0
zeromq=4.3.3=he6710b0_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h6597ccf_2
